@@ -6,15 +6,16 @@ buildscript {
             url "https://plugins.gradle.org/m2/"
         }
     }
+
     dependencies {
-        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.1.1.RELEASE")
+        classpath("org.springframework.boot:spring-boot-gradle-plugin:2.6.6")
     }
 }
 
 plugins {
     id 'java'
     id "com.moowork.node" version "1.2.0"
-    id "org.asciidoctor.convert" version "1.5.3"
+    id 'org.asciidoctor.jvm.convert' version '3.3.2'
 }
 
 allprojects {
@@ -35,6 +36,7 @@ apply plugin: 'io.spring.dependency-management'
 apply plugin: "com.moowork.node"
 
 node {
+//    version = '14.18.0'
     version = '10.14.1'
     download = true
     // Set the work directory for unpacking node
@@ -46,10 +48,10 @@ node {
 }
 
 bootJar {
-    baseName = 'gs-spring-boot'
-    version =  '0.1.0'
+    archiveBaseName = 'gs-spring-boot'
+    archiveVersion = '0.1.0'
     dependsOn asciidoctor
-    from ("${asciidoctor.outputDir}/html5") {
+    from("${asciidoctor.outputDir}/html5") {
         into 'static/docs'
     }
 }
@@ -57,33 +59,36 @@ bootJar {
 group 'org.haffson'
 version '1.0-SNAPSHOT'
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 17
+targetCompatibility = 17
 
 ext {
     snippetsDir = file('target/snippets')
 }
 
 dependencies {
+    implementation 'org.springframework.boot:spring-boot:2.6.6'
     implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.boot:spring-boot-starter-hateoas'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'com.h2database:h2'
-    implementation 'org.postgresql:postgresql:42.2.1'
+    implementation 'org.postgresql:postgresql:42.2.24'
     implementation 'org.liquibase:liquibase-core'
-    implementation 'com.google.code.findbugs:jsr305:1.3.9'
-    asciidoctor 'org.springframework.restdocs:spring-restdocs-asciidoctor'
+    implementation 'com.google.code.findbugs:jsr305:3.0.2'
+    implementation 'org.apache.commons:commons-lang3:3.12.0'
     testImplementation("org.springframework.restdocs:spring-restdocs-mockmvc")
+    testImplementation 'org.mockito:mockito-inline:4.6.1'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'junit:junit'
-    testImplementation 'org.mockito:mockito-inline:2.8.9'
-    testImplementation 'com.tngtech.archunit:archunit:0.9.3'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
+    testImplementation 'com.tngtech.archunit:archunit:0.21.0'
 }
 
 test {
     outputs.dir snippetsDir
+    useJUnitPlatform()
 }
 
 asciidoctor {
@@ -116,7 +121,7 @@ tasks.withType(Test) {
     }
 }
 
-if(Boolean.valueOf(System.getenv("FE_BUILD"))) {
+if (Boolean.valueOf(System.getenv("FE_BUILD"))) {
     task npmRunScriptBuild(type: NpmTask, dependsOn: 'npmTest') {
         args = ['run-script', 'build:jar']
         execOverrides {
@@ -133,9 +138,9 @@ if(Boolean.valueOf(System.getenv("FE_BUILD"))) {
     npmTest.dependsOn npmInstall
     npmRunScriptBuild.dependsOn npmTest
 
-/** processResources is part of the jar task and copies production resources into the production
- *  resources directory. This should only happen after then npmRunScriptBuild has generated those
- *  resources.
- */
+    /** processResources is part of the jar task and copies production resources into the production
+     *  resources directory. This should only happen after then npmRunScriptBuild has generated those
+     *  resources.
+     */
     processResources.dependsOn npmRunScriptBuild
 }
