@@ -5,12 +5,12 @@ import com.tngtech.archunit.core.importer.ClassFileImporter;
 import com.tngtech.archunit.library.GeneralCodingRules;
 import org.haffson.adventofcode.archUnitTestingUtils.MethodCondition;
 import org.haffson.adventofcode.archUnitTestingUtils.MethodParameterNumberPredicate;
-import org.haffson.adventofcode.archUnitTestingUtils.MethodTransformer;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static com.tngtech.archunit.lang.conditions.ArchConditions.callMethod;
-import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.no;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses;
+import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noMethods;
+
 
 public class AntiPatternDetectionTest {
 
@@ -33,9 +33,10 @@ public class AntiPatternDetectionTest {
 
     @Test
     public void doNotUseTooManyMethodParameters() {
-        no(MethodTransformer.methods())
+        noMethods()
                 .should(MethodCondition.have(MethodParameterNumberPredicate.moreParametersThan(4)))
                 .because("methods with too many parameters are hard to understand and complicated to test. Consider refactoring, e.g. splitting functionality or introducing Parameter Objects.")
                 .check(allClasses);
     }
+
 }
