@@ -3,7 +3,7 @@ package org.haffson.adventofcode;
 import com.tngtech.archunit.core.domain.JavaClasses;
 import com.tngtech.archunit.core.importer.ClassFileImporter;
 import com.tngtech.archunit.core.importer.ImportOption;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.net.URL;
 
@@ -17,13 +17,14 @@ public class ArchUnitRulesFromPlantUML {
     URL myDiagram = getClass().getClassLoader().getResource("adventOfCodeArchitecture.puml");
 
     private final JavaClasses allClasses = new ClassFileImporter()
-            .withImportOption(new ImportOption.DontIncludeTests())
+            .withImportOption(new ImportOption.DoNotIncludeTests())
             .importPackages("org.haffson.adventofcode");
 
     @Test
     public void classesShouldAdhereToPlantUmlDiagram() {
         classes().should(adhereToPlantUmlDiagram(myDiagram, consideringOnlyDependenciesInDiagram()))
-                .check(allClasses);;
+                .check(allClasses);
+        ;
     }
 
 }
