@@ -2,9 +2,8 @@ package org.haffson.adventofcode.controller;
 
 import org.haffson.adventofcode.days.Days;
 import org.haffson.adventofcode.service.AdventOfCodeService;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
@@ -12,12 +11,12 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.restdocs.payload.FieldDescriptor;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.io.FileNotFoundException;
-import java.nio.charset.StandardCharsets;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.LinkedList;
 import java.util.List;
 
 import static org.hamcrest.Matchers.is;
@@ -30,39 +29,54 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@RunWith(SpringRunner.class)
 @WebMvcTest(AdventOfCodeController.class)
 @AutoConfigureRestDocs(outputDir = "target/snippets")
-public class AdventOfCodeControllerTest {
+class AdventOfCodeControllerTest {
 
     @MockBean
-    private AdventOfCodeService adventOfCodeService;
+    AdventOfCodeService adventOfCodeService;
 
     private final String baseUrl = "/api/adventOfCode";
 
-    private final MediaType contentType = new MediaType("application", "hal+json", StandardCharsets.UTF_8);
+    private final MediaType contentType = new MediaType("application", "hal+json", Charset.forName("UTF-8"));
 
-    private final String day1 = "1";
-    private final String part1 = "1";
-    private final String resultDay1Part1 = "Part 1 - Frequency: 599";
+    private final Integer day1 = 1;
+    private final Integer part1 = 1;
+    private final Integer part2 = 2;
+    private final String resultDay1Part1 = "Product 1: " + 326211;
+    private final String resultDay1Part2 = "Product 2: " + 326211;
 
     @Autowired
     private MockMvc mvc;
 
-    @Before
-    public void setup() throws FileNotFoundException {
+
+    @BeforeEach
+    void setup() {
+        final Days day01Stub = Mockito.mock(Days.class);
+        final Days day02Stub = Mockito.mock(Days.class);
+        Mockito.when(day01Stub.getDay()).thenReturn(1);
+        Mockito.when(day02Stub.getDay()).thenReturn(2);
+
+        final List<Days> daysImplementedList = new LinkedList<>();
+        daysImplementedList.add(day01Stub);
+        daysImplementedList.add(day02Stub);
+
         Mockito.when(adventOfCodeService.getResultsForASpecificDayAndPuzzlePart(day1, part1))
                 .thenReturn(resultDay1Part1);
+        Mockito.when(adventOfCodeService.getDaysSolutions())
+                .thenReturn(daysImplementedList);
     }
 
-    @Test
-    public void testGetResultForASpecificDayAndPuzzlePart() throws Exception {
 
-        mvc.perform(get(baseUrl + "?day=" + day1 + "&part=" + part1)
+    @Test
+    void testGetResultForASpecificDayAndPuzzlePart() throws Exception {
+        mvc.perform(get(baseUrl + "/" + "?day=" + day1 + "&part=" + part1)
                 .contentType(contentType))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("content", is(resultDay1Part1)))
-                .andExpect(jsonPath("$._links.self.href", is("http://localhost:8080" + baseUrl + "?day=" + day1 + "&part=" + part1)))
+                .andExpect(jsonPath("day", is(1)))
+                .andExpect(jsonPath("part", is(1)))
+                .andExpect(jsonPath("answer", is("Product 1: " + 326211)))
+                .andExpect(jsonPath("$._links.self.href", is("http://localhost:8080" + baseUrl + "/" + "?day=" + day1 + "&part=" + part1)))
                 .andDo(document("getResultForASpecificDayAndPuzzlePart",
                         preprocessResponse(prettyPrint()),
                         responseFields(getResultForASpecificDayAndPuzzlePart("")))
@@ -70,28 +84,30 @@ public class AdventOfCodeControllerTest {
     }
 
     @Test
-    public void testDaysImplementedReturnsSortedList() throws Exception {
-        Days day01Stub = Mockito.mock(Days.class);
-        Mockito.when(day01Stub.getDay()).thenReturn(1);
+    void testGetResultForASpecificDayAndPuzzlePart2() throws Exception {
+        mvc.perform(get(baseUrl + "/" + "?day=" + day1 + "&part=" + part2)
+                .contentType(contentType))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("day", is(1)))
+                .andExpect(jsonPath("part", is(2)))
+//                .andExpect(jsonPath("answer", is("Product 1: " + 326211)))
+                .andExpect(jsonPath("$._links.self.href", is("http://localhost:8080" + baseUrl + "/" + "?day=" + day1 + "&part=" + part2)))
+                .andDo(document("getResultForASpecificDayAndPuzzlePart",
+                        preprocessResponse(prettyPrint()),
+                        responseFields(getResultForASpecificDayAndPuzzlePart("")))
+                );
 
-        Days day02Stub = Mockito.mock(Days.class);
-        Mockito.when(day02Stub.getDay()).thenReturn(2);
+    }
 
-        List<Days> daysImplementedList = new ArrayList<>();
-        daysImplementedList.add(day02Stub);
-        daysImplementedList.add(day01Stub);
+    @Test
+    void testDaysImplemented() throws Exception {
 
-        List<Integer> daysImplementedIntegerList = new ArrayList<>();
-        daysImplementedIntegerList.add(1);
-        daysImplementedIntegerList.add(2);
-
-        Mockito.when(adventOfCodeService.getDaysSolutions())
-                .thenReturn(daysImplementedList);
+        final List<Integer> daysImplementedIntegerList = new LinkedList<>(Arrays.asList(1, 2));
 
         mvc.perform(get(baseUrl + "/daysimplemented")
                 .contentType(contentType))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("_embedded.integerList", is(daysImplementedIntegerList)))
+                .andExpect(jsonPath("$._embedded.integerList", is(daysImplementedIntegerList)))
                 .andExpect(jsonPath("$._links.self.href", is("http://localhost:8080" + baseUrl + "/daysimplemented")))
                 .andDo(document(
                         "daysImplemented",
@@ -101,14 +117,18 @@ public class AdventOfCodeControllerTest {
                 );
     }
 
-    private ArrayList<FieldDescriptor> getResultForASpecificDayAndPuzzlePart(String path) {
-        String pathString;
+    private ArrayList<FieldDescriptor> getResultForASpecificDayAndPuzzlePart(final String path) {
+        final String pathString;
         if (path.isEmpty()) {
             pathString = "";
         } else pathString = path;
 
-        ArrayList<FieldDescriptor> fieldDescriptorList = new ArrayList<>();
-        fieldDescriptorList.add(fieldWithPath(pathString + "content")
+        final ArrayList<FieldDescriptor> fieldDescriptorList = new ArrayList<>();
+        fieldDescriptorList.add(fieldWithPath(pathString + "day")
+                .description("Specific day of the puzzle of the AdventOfCode calendar"));
+        fieldDescriptorList.add(fieldWithPath(pathString + "part")
+                .description("Specific day's part of the puzzle of the AdventOfCode calendar"));
+        fieldDescriptorList.add(fieldWithPath(pathString + "answer")
                 .description("Result of the Puzzle for a specific day and part of the AdventOfCode calendar"));
         fieldDescriptorList.add(fieldWithPath(pathString + "_links.self.href")
                 .description("Self link to the query for the specific solution for a day and part"));
@@ -116,13 +136,13 @@ public class AdventOfCodeControllerTest {
         return fieldDescriptorList;
     }
 
-    private ArrayList<FieldDescriptor> daysImplemented(String path) {
-        String pathString;
+    private ArrayList<FieldDescriptor> daysImplemented(final String path) {
+        final String pathString;
         if (path.isEmpty()) {
             pathString = "";
         } else pathString = path;
 
-        ArrayList<FieldDescriptor> fieldDescriptorList = new ArrayList<>();
+        final ArrayList<FieldDescriptor> fieldDescriptorList = new ArrayList<>();
         fieldDescriptorList.add(fieldWithPath(pathString + "_embedded.integerList")
                 .description("List of all implemented days"));
         fieldDescriptorList.add(fieldWithPath(pathString + "_links.self.href")
