@@ -1,12 +1,9 @@
 package org.haffson.adventofcode.days.day03;
 
-import org.haffson.adventofcode.days.day01.Day01;
 import org.haffson.adventofcode.utils.FileReaders;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit4.SpringRunner;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -15,22 +12,22 @@ import java.util.List;
 
 import static org.mockito.Mockito.when;
 
-@RunWith(SpringRunner.class)
-public class Day03Test {
+
+class Day03Test {
 
     @MockBean
     private FileReaders fileReaders;
 
     @Test
-    public void testGetDay() {
+    void testGetDay() {
         final Day03 day03 = new Day03(fileReaders);
         final int expectedResult = 3;
         final int actualResult = day03.getDay();
-        Assert.assertEquals(expectedResult, actualResult);
+        assertEquals(expectedResult, actualResult);
     }
 
     @Test
-    public void test_firstPart_returnsExpectedResult() {
+    void test_firstPart_returnsExpectedResult() {
         //arrange
         final Day03 day03 = new Day03(fileReaders);
         when(fileReaders.getInputList("src/main/resources/puzzle_input/day3_input.txt"))
@@ -48,11 +45,11 @@ public class Day03Test {
         final String actualResult = day03.firstPart();
 
         //assert
-        Assert.assertEquals(expectedResult, actualResult);
+        assertEquals(expectedResult, actualResult);
     }
 
     @Test
-    public void testConvertStringToRectangleList_returnListOfRectangles() {
+    void testConvertStringToRectangleList_returnListOfRectangles() {
         //arrange
         final Day03 day03 = new Day03(fileReaders);
         final List<String> stringList = Arrays.asList(
@@ -70,11 +67,11 @@ public class Day03Test {
         final List<RectangleClaim> actualResult = day03.convertStringToRectangleList(stringList);
 
         //assert
-        Assert.assertEquals(expectedResult, actualResult);
+        assertEquals(expectedResult, actualResult);
     }
 
     @Test
-    public void test_secondPart_returnsExpectedResult() {
+    void test_secondPart_returnsExpectedResult() {
         //arrange
         final Day03 day03 = new Day03(fileReaders);
         when(fileReaders.getInputList("src/main/resources/puzzle_input/day3_input.txt"))
@@ -91,6 +88,6 @@ public class Day03Test {
         final String actualResult = day03.secondPart();
 
         //assert
-        Assert.assertEquals(expectedResult, actualResult);
+        assertEquals(expectedResult, actualResult);
     }
 }
