@@ -1,11 +1,12 @@
 package org.haffson.adventofcode.utils;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+//import org.junit.Test;
+import org.junit.jupiter.api.Test;
+//import org.junit.runner.RunWith;
 import org.springframework.test.context.junit4.SpringRunner;
 
 
-@RunWith(SpringRunner.class)
+//@RunWith(SpringRunner.class)
 public class FileReadersTest {
 
     //TODO: Meaningful tests for readFileIntoArrayOfIntegers and readFileIntoStringList
