@@ -37,7 +37,7 @@ const styles = {
 function renderAddNewCardActions(cardType, classes, callController) {
   if (cardType === "newCard") {
     return (
-      <CardActions className={classes.actions} disableActionSpacing>
+      <CardActions className={classes.actions} disableactionspacing="true">
         <IconButton onClick={callController}>
           <FloatingActionButtons />
         </IconButton>
@@ -107,11 +107,11 @@ class CardTemplate extends Component {
   async callController() {
     this.setState({ loading: true });
     const response = await fetch(
-      "/api/adventOfCode?day=" + this.state.day + "&part=" + this.state.value
+      "/api/adventOfCode/?day=" + this.state.day + "&part=" + this.state.value
     );
     try {
       const data = await response.json();
-      const content = data.content;
+      const content = data.answer;
       const message = data.message;
       if (response.ok) {
         this.props.callback({
