@@ -6,12 +6,18 @@ package org.haffson.adventofcode;
  * @author Michelle Fernandez Bieber
  */
 public enum ProblemStatusEnum {
-    /** if the puzzle has been solved */
+    /**
+     * if the puzzle has been solved
+     */
     SOLVED,
 
-    /** if the puzzle is being worked on */
+    /**
+     * if the puzzle is being worked on
+     */
     IN_PROGRESS,
 
-    /** if the puzzle hasn't been started yet */
+    /**
+     * if the puzzle hasn't been started yet
+     */
     UNSOLVED
 }
